Code/backend/requirements.txt
bcrypt==4.0.1
blinker==1.7.0
bleach == 6.1.0
click==8.1.7
colorama==0.4.6
Flask== 3.0.0
Flask-Caching==2.1.0
Flask-Cors==3.0.10
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.28.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
pytest == 7.4.3
Flask-Security-Too == 5.3.2
matplotlib ==3.8.2



